Update tests Euler 2D
Switch to pytest and use a single test file.
I am not using the tools from pyclaw.utils. Should we test them with
the examples? If so, maybe we should keep them in one or two examples.
I am skipping tests for shock_forward_step, so the HLLE RS with walls is
not tested. This example gives me some random NaNs that I am not able to
handle robustly.