Update test Euler with gravity 3d
I didn't see why the test (with nose) was skipped
if scipy couldn't be found. I removed that requirement here.

The execution of both examples is not aborted now if mpi4py
is not available, just a warning is displayed.

The same .txt file with
data for the regression test is used.

The local Makefile is used to
compile the source with subprocess.